Full removal of tasks and dataclasses
