+++ recirq/time_crystals/__init__.py
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from recirq.time_crystals.dtctask import DTCTask, CompareDTCTask
+from recirq.time_crystals.dtcexperiment import (DTCExperiment, comparison_experiments,
+                                        EXPERIMENT_NAME, DEFAULT_BASE_DIR)
+
